move: data 모듈 리패키징
