--- data/src/main/java/com/susu/data/Constants.kt
@@ -1,5 +1,0 @@
-package com.susu.data
-
-object Constants {
-    const val RETROFIT_TAG = "Retrofit"
-}
