--- data/src/main/java/com/susu/data/Json.kt
@@ -1,10 +1,0 @@
-package com.susu.data
-
-import com.susu.data.retrofit.SusuErrorResponse
-import kotlinx.serialization.json.Json
-
-object Json {
-    val instance = Json { encodeDefaults = true }
-
-    fun getSusuErrorBody(body: String) = instance.decodeFromString<SusuErrorResponse>(body)
-}
