--- data/src/main/java/com/susu/data/di/Qualifier.kt
@@ -1,11 +1,0 @@
-package com.susu.data.di
-
-import javax.inject.Qualifier
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-internal annotation class AuthOkHttpClient
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-internal annotation class AuthRetrofit
