--- data/src/main/java/com/susu/data/extension/StringExtension.kt
@@ -1,8 +1,0 @@
-package com.susu.data.extension
-
-import org.json.JSONArray
-import org.json.JSONObject
-
-fun String.isJsonObject(): Boolean = runCatching { JSONObject(this) }.isSuccess
-
-fun String.isJsonArray(): Boolean = runCatching { JSONArray(this) }.isSuccess
