+++ data/src/main/java/com/susu/data/local/model/EntityTable.kt
@@ -1,0 +1,6 @@
+package com.susu.data.local.model
+
+object EntityTable {
+    const val LEDGER_RECENT_SEARCH = "LedgerRecentSearch"
+    const val ENVELOPE_RECENT_SEARCH = "EnvelopeRecentSearch"
+}
