+++ data/src/main/java/com/susu/data/local/model/EnvelopeRecentSearchEntity.kt
@@ -1,0 +1,11 @@
+package com.susu.data.local.model
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = EntityTable.ENVELOPE_RECENT_SEARCH)
+data class EnvelopeRecentSearchEntity(
+    @PrimaryKey
+    val search: String,
+    val saveTime: Long,
+)
