--- data/src/main/java/com/susu/data/model/SnsProviders.kt
@@ -1,5 +1,0 @@
-package com.susu.data.model
-
-enum class SnsProviders(val path: String) {
-    Kakao("KAKAO"),
-}
