--- data/src/main/java/com/susu/data/model/entity/EntityTable.kt
@@ -1,6 +1,0 @@
-package com.susu.data.model.entity
-
-object EntityTable {
-    const val LEDGER_RECENT_SEARCH = "LedgerRecentSearch"
-    const val ENVELOPE_RECENT_SEARCH = "EnvelopeRecentSearch"
-}
