--- data/src/main/java/com/susu/data/model/request/AccessTokenRequest.kt
@@ -1,8 +1,0 @@
-package com.susu.data.model.request
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class AccessTokenRequest(
-    val accessToken: String,
-)
