--- data/src/main/java/com/susu/data/model/request/RefreshTokenRequest.kt
@@ -1,9 +1,0 @@
-package com.susu.data.model.request
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RefreshTokenRequest(
-    val accessToken: String,
-    val refreshToken: String,
-)
