--- data/src/main/java/com/susu/data/model/request/UserRequest.kt
@@ -1,17 +1,0 @@
-package com.susu.data.model.request
-
-import com.susu.core.model.User
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class UserRequest(
-    val name: String,
-    val gender: String,
-    val birth: Int,
-)
-
-fun User.toData() = UserRequest(
-    name = name,
-    gender = gender,
-    birth = birth,
-)
