--- data/src/main/java/com/susu/data/model/response/ValidRegisterResponse.kt
@@ -1,8 +1,0 @@
-package com.susu.data.model.response
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ValidRegisterResponse(
-    val canRegister: Boolean,
-)
