+++ data/src/main/java/com/susu/data/remote/Constants.kt
@@ -1,0 +1,5 @@
+package com.susu.data.remote
+
+object Constants {
+    const val RETROFIT_TAG = "Retrofit"
+}
