+++ data/src/main/java/com/susu/data/remote/Json.kt
@@ -1,0 +1,10 @@
+package com.susu.data.remote
+
+import com.susu.data.remote.retrofit.SusuErrorResponse
+import kotlinx.serialization.json.Json
+
+object Json {
+    val instance = Json { encodeDefaults = true }
+
+    fun getSusuErrorBody(body: String) = instance.decodeFromString<SusuErrorResponse>(body)
+}
