+++ data/src/main/java/com/susu/data/remote/di/Qualifier.kt
@@ -1,0 +1,11 @@
+package com.susu.data.remote.di
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+internal annotation class AuthOkHttpClient
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+internal annotation class AuthRetrofit
