+++ data/src/main/java/com/susu/data/remote/model/SnsProviders.kt
@@ -1,0 +1,5 @@
+package com.susu.data.remote.model
+
+enum class SnsProviders(val path: String) {
+    Kakao("KAKAO"),
+}
