+++ data/src/main/java/com/susu/data/remote/model/request/AccessTokenRequest.kt
@@ -1,0 +1,8 @@
+package com.susu.data.remote.model.request
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AccessTokenRequest(
+    val accessToken: String,
+)
