+++ data/src/main/java/com/susu/data/remote/model/request/RefreshTokenRequest.kt
@@ -1,0 +1,9 @@
+package com.susu.data.remote.model.request
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RefreshTokenRequest(
+    val accessToken: String,
+    val refreshToken: String,
+)
