+++ data/src/main/java/com/susu/data/remote/model/response/TokenResponse.kt
@@ -1,0 +1,15 @@
+package com.susu.data.remote.model.response
+
+import com.susu.core.model.Token
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TokenResponse(
+    val accessToken: String,
+    val refreshToken: String,
+)
+
+fun TokenResponse.toDomain() = Token(
+    accessToken = accessToken,
+    refreshToken = refreshToken,
+)
