+++ data/src/main/java/com/susu/data/remote/model/response/ValidRegisterResponse.kt
@@ -1,0 +1,8 @@
+package com.susu.data.remote.model.response
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ValidRegisterResponse(
+    val canRegister: Boolean,
+)
