+++ data/src/main/java/com/susu/data/remote/retrofit/SusuErrorResponse.kt
@@ -1,0 +1,10 @@
+package com.susu.data.remote.retrofit
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SusuErrorResponse(
+    val errorCode: String,
+    val reason: String,
+    val extra: Map<String, String>? = null,
+)
