+++ data/src/main/java/com/susu/data/repository/SignUpRepositoryImpl.kt
@@ -1,10 +1,10 @@
 package com.susu.data.repository
 
 import com.susu.core.model.User
-import com.susu.data.model.SnsProviders
-import com.susu.data.model.request.toData
-import com.susu.data.model.response.toDomain
-import com.susu.data.network.service.SignUpService
+import com.susu.data.remote.model.SnsProviders
+import com.susu.data.remote.model.request.toData
+import com.susu.data.remote.model.response.toDomain
+import com.susu.data.remote.api.SignUpService
 import com.susu.domain.repository.SignUpRepository
 import javax.inject.Inject
