--- data/src/main/java/com/susu/data/retrofit/SusuErrorResponse.kt
@@ -1,10 +1,0 @@
-package com.susu.data.retrofit
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SusuErrorResponse(
-    val errorCode: String,
-    val reason: String,
-    val extra: Map<String, String>? = null,
-)
